+++ react/__mocks__/context.ts
@@ -1,0 +1,10 @@
+const contextValuesMock = {
+  zipcode: '',
+  selectedItem: {},
+  selectedQuantity: '0',
+  pickupSlas: [],
+  setZipcode: () => {},
+  setPickupSlas: () => {},
+}
+
+export default contextValuesMock