+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Pickup point list block
+- Pickup point option card block
+- Pickup point name block
+- Tests for all the blocks added
+
 ## [0.1.1] - 2021-09-17
 
 ### Added